deleted redundant Maps activities
