+++ app/build.gradle
@@ -45,7 +45,8 @@ dependencies {
 
     // Dependency to include Maps SDK for Android
     implementation 'com.google.android.gms:play-services-maps:17.0.0'
-
+    //Dependency to include Places SDK for Android
+    implementation 'com.google.android.libraries.places:places:2.4.0'
     implementation 'androidx.appcompat:appcompat:1.2.0'
     implementation 'com.google.android.material:material:1.3.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
