+++ app/src/main/AndroidManifest.xml
@@ -30,9 +30,7 @@
             android:name="com.google.android.geo.API_KEY"
             android:value="@string/google_maps_key" />
 
-        <activity
-            android:name=".MapsActivity"
-            android:label="@string/title_activity_maps"></activity>
+
         <activity
             android:name=".MapActivity"
             android:label="@string/title_map" />
