+++ build.gradle
@@ -17,6 +17,9 @@ buildscript {
 
 allprojects {
     repositories {
+        maven {
+            url 'https://maven.google.com'
+        }
         google()
         jcenter()
     }
